Finished UCMA bot implementation
Finished UCMA bot implementation with support for Lync meeting
transcripts on incoming Audio or InstantMessaing calls, Conference
invitations, and/or established calls that are escalated to conferences.
Also added logic to handle subscribing to events on sub conversations
for Calls.

TODO: Clean up code a bit and add meeting transcript messages for all
events subscribed to on calls, conversations, and conference sessions.